Extract documentation rendering
