+++ compiler-bin/src/docs.rs
@@ -1,22 +1,16 @@
 pub mod error;
-pub mod schema;
-pub mod warm;
 
 use std::collections::BTreeMap;
 use std::path::{Path, PathBuf};
 use std::{env, fs, process};
 
 use analyzer::{QueryEngine, prim};
-use checking::PrettyQueries;
-use checking::core::pretty::PrettyNames;
 use files::{FileId, Files};
 use itertools::Itertools;
 use rayon::iter::ParallelIterator;
 use serde::Deserialize;
-use ts_rs::{Config as TypeScriptExportConfig, TS};
 
 use crate::docs::error::DocsError;
-use crate::docs::warm::warm_documentation_queries;
 use crate::walk;
 
 pub struct DocsConfig {
@@ -46,11 +40,7 @@ pub fn typescript(config: TypeScriptConfig) {
 }
 
 fn write_typescript(config: TypeScriptConfig) -> Result<(), DocsError> {
-    let typescript = TypeScriptExportConfig::new().with_out_dir(config.output);
-
-    schema::Package::export_all(&typescript)?;
-    schema::Module::export_all(&typescript)?;
-
+    documentation::export_typescript(config.output)?;
     Ok(())
 }
 
@@ -95,131 +85,6 @@ struct PursManifest {
     dependencies: BTreeMap<String, String>,
 }
 
-struct TypeEncoder<'a> {
-    engine: &'a QueryEngine,
-    checked: &'a checking::CheckedModule,
-    package_by_file: &'a [(FileId, &'a str)],
-    names: PrettyNames,
-}
-
-impl<'a> TypeEncoder<'a> {
-    fn new(
-        engine: &'a QueryEngine,
-        checked: &'a checking::CheckedModule,
-        package_by_file: &'a [(FileId, &'a str)],
-    ) -> TypeEncoder<'a> {
-        TypeEncoder { engine, checked, package_by_file, names: PrettyNames::new() }
-    }
-
-    fn encode_signature(&mut self, id: checking::TypeId) -> Result<schema::Type, DocsError> {
-        self.names.reset();
-        self.encode_type(id)
-    }
-
-    fn encode_type(&mut self, id: checking::TypeId) -> Result<schema::Type, DocsError> {
-        let expression = match self.engine.lookup_type(id) {
-            checking::Type::Application(function, argument) => schema::Type::Application {
-                function: self.encode_boxed_type(function)?,
-                argument: self.encode_boxed_type(argument)?,
-            },
-            checking::Type::KindApplication(function, argument) => schema::Type::KindApplication {
-                function: self.encode_boxed_type(function)?,
-                argument: self.encode_boxed_type(argument)?,
-            },
-            checking::Type::Forall(binder, body) => schema::Type::Forall {
-                binder: self.encode_binder(binder)?,
-                body: self.encode_boxed_type(body)?,
-            },
-            checking::Type::Constrained(constraint, body) => schema::Type::Constrained {
-                constraint: self.encode_boxed_type(constraint)?,
-                body: self.encode_boxed_type(body)?,
-            },
-            checking::Type::Function(argument, result) => schema::Type::Function {
-                argument: self.encode_boxed_type(argument)?,
-                result: self.encode_boxed_type(result)?,
-            },
-            checking::Type::Kinded(expression, kind) => schema::Type::Kinded {
-                expression: self.encode_boxed_type(expression)?,
-                kind: self.encode_boxed_type(kind)?,
-            },
-            checking::Type::Constructor(file_id, type_id) => schema::Type::Constructor {
-                reference: self.resolve_type_reference(file_id, type_id)?,
-            },
-            checking::Type::Integer(value) => schema::Type::Integer { value },
-            checking::Type::String(kind, value_id) => {
-                let kind = match kind {
-                    lowering::StringKind::String => schema::StringLiteralKind::String,
-                    lowering::StringKind::RawString => schema::StringLiteralKind::RawString,
-                };
-                let value = self.engine.lookup_smol_str(value_id).to_string();
-                schema::Type::String { kind, value }
-            }
-            checking::Type::Row(row_id) => {
-                let row = self.engine.lookup_row_type(row_id);
-                let fields = row.fields.iter().map(|field| {
-                    let t = self.encode_type(field.id)?;
-                    Ok(schema::TypeRowField { label: field.label.to_string(), t })
-                });
-
-                let fields = fields.collect::<Result<Vec<_>, DocsError>>()?;
-                let tail = row.tail.map(|id| self.encode_boxed_type(id)).transpose()?;
-
-                schema::Type::Row { fields, tail }
-            }
-            checking::Type::Rigid(name, _, kind) => schema::Type::Rigid {
-                name: self.display_name(name),
-                kind: self.encode_boxed_type(kind)?,
-            },
-            checking::Type::Unification(id) => schema::Type::Unification { id },
-            checking::Type::Free(name_id) => {
-                schema::Type::Free { name: self.engine.lookup_smol_str(name_id).to_string() }
-            }
-            checking::Type::Unknown(name_id) => {
-                schema::Type::Unknown { name: self.engine.lookup_smol_str(name_id).to_string() }
-            }
-        };
-
-        Ok(expression)
-    }
-
-    fn encode_boxed_type(&mut self, id: checking::TypeId) -> Result<Box<schema::Type>, DocsError> {
-        Ok(Box::new(self.encode_type(id)?))
-    }
-
-    fn encode_binder(
-        &mut self,
-        id: checking::core::ForallBinderId,
-    ) -> Result<schema::TypeBinder, DocsError> {
-        let binder = self.engine.lookup_forall_binder(id);
-        let name = self.display_name(binder.name);
-        let kind = self.encode_boxed_type(binder.kind)?;
-
-        Ok(schema::TypeBinder { name, visible: binder.visible, kind })
-    }
-
-    fn resolve_type_reference(
-        &self,
-        file_id: FileId,
-        type_id: indexing::TypeItemId,
-    ) -> Result<schema::TypeReference, DocsError> {
-        let package = self.package_by_file.iter().find_map(|&(id, package)| {
-            if id == file_id { Some(package.to_string()) } else { None }
-        });
-
-        let (parsed, _) = self.engine.parsed(file_id)?;
-        let module = parsed.module_name().map(|name| name.to_string());
-
-        let indexed = self.engine.indexed(file_id)?;
-        let name = indexed.items[type_id].name.as_ref().map(|name| name.to_string());
-
-        Ok(schema::TypeReference { package, module, name })
-    }
-
-    fn display_name(&mut self, name: checking::core::Name) -> String {
-        self.names.display_name(self.engine, &self.checked.names, name).to_string()
-    }
-}
-
 fn generate_documentation(config: DocsConfig) -> Result<(), DocsError> {
     let mut compiler = Compiler::default();
     prim::configure(&mut compiler.engine, &mut compiler.files);
@@ -227,7 +92,7 @@ fn generate_documentation(config: DocsConfig) -> Result<(), DocsError> {
     let packages = load_packages(&config, &mut compiler)?;
     let modules = package_modules(&packages);
 
-    warm_documentation_queries(&compiler.engine, &modules)?;
+    documentation::warm_queries(&compiler.engine, &modules)?;
     write_packages_manifest(&compiler.engine, &config, &packages)?;
 
     let package_by_file = packages
@@ -389,21 +254,15 @@ fn write_packages_manifest(
     let root = env::current_dir()?;
 
     for package in packages {
-        let modules = package.modules.iter().map(|&id| {
-            let (parsed, _) = engine.parsed(id)?;
-            Ok(parsed.module_name().map(|name| name.to_string()))
-        });
-
-        let modules = modules.collect::<Result<Vec<_>, DocsError>>()?;
-
-        let package = schema::Package {
+        let package_input = documentation::PackageInput {
             name: String::clone(&package.name),
             version: String::clone(&package.version),
             license: package.license.clone(),
             description: package.description.clone(),
             dependencies: BTreeMap::clone(&package.dependencies),
-            modules: modules.into_iter().flatten().collect_vec(),
+            modules: package.modules.clone(),
         };
+        let package = documentation::render_package_manifest(engine, &package_input)?;
 
         let package_folder = root.join(&config.output).join(&package.name);
         fs::create_dir_all(&package_folder)?;
@@ -428,67 +287,12 @@ fn generate_package_documentation(
     fs::create_dir_all(&modules_folder)?;
 
     for &id in &package.modules {
-        let (parsed, _) = compiler.engine.parsed(id)?;
-        let indexed = compiler.engine.indexed(id)?;
-        let checked = compiler.engine.checked(id)?;
-        let documented = compiler.engine.documented(id)?;
-
-        let Some(name) = parsed.module_name().map(|name| name.to_string()) else {
+        let Some(module) = documentation::render_module(&compiler.engine, id, package_by_file)?
+        else {
             continue;
         };
 
-        let mut terms = vec![];
-        let mut types = vec![];
-        let mut type_encoder = TypeEncoder::new(&compiler.engine, &checked, package_by_file);
-
-        for (term_id, term_item) in indexed.items.iter_terms() {
-            let term_documentation = documented.terms.get(&term_id);
-
-            let name = term_item.name.as_ref().map(|name| name.to_string());
-            let documentation = term_documentation.map(|t| t.documentation.to_string());
-            let signature = checked
-                .lookup_term(term_id)
-                .map(|signature| type_encoder.encode_signature(signature))
-                .transpose()?;
-
-            let kind = match &term_item.kind {
-                indexing::TermItemKind::ClassMember { .. } => schema::TermKind::ClassMember,
-                indexing::TermItemKind::Constructor { .. } => schema::TermKind::Constructor,
-                indexing::TermItemKind::Derive { .. } => schema::TermKind::Derive,
-                indexing::TermItemKind::Foreign { .. } => schema::TermKind::Foreign,
-                indexing::TermItemKind::Instance { .. } => schema::TermKind::Instance,
-                indexing::TermItemKind::Operator { .. } => schema::TermKind::Operator,
-                indexing::TermItemKind::Value { .. } => schema::TermKind::Value,
-            };
-
-            terms.push(schema::TermItem { name, documentation, signature, kind });
-        }
-
-        for (type_id, type_item) in indexed.items.iter_types() {
-            let type_documentation = documented.types.get(&type_id);
-
-            let name = type_item.name.as_ref().map(|name| name.to_string());
-            let documentation = type_documentation.map(|t| t.documentation.to_string());
-            let signature = checked
-                .lookup_type(type_id)
-                .map(|signature| type_encoder.encode_signature(signature))
-                .transpose()?;
-
-            let kind = match &type_item.kind {
-                indexing::TypeItemKind::Data { .. } => schema::TypeKind::Data,
-                indexing::TypeItemKind::Newtype { .. } => schema::TypeKind::Newtype,
-                indexing::TypeItemKind::Synonym { .. } => schema::TypeKind::Synonym,
-                indexing::TypeItemKind::Class { .. } => schema::TypeKind::Class,
-                indexing::TypeItemKind::Foreign { .. } => schema::TypeKind::Foreign,
-                indexing::TypeItemKind::Operator { .. } => schema::TypeKind::Operator,
-            };
-
-            types.push(schema::TypeItem { name, documentation, signature, kind });
-        }
-
-        let module_file = modules_folder.join(format!("{name}.json"));
-        let documentation = Some(documented.documentation.to_string());
-        let module = schema::Module { name, documentation, terms, types };
+        let module_file = modules_folder.join(format!("{}.json", module.name));
         let module = serde_json::to_string_pretty(&module)?;
 
         fs::write(module_file, module)?;
